requirements/base.txt
# Base
pytz==2021.1
python-slugify==1.2.6
Pillow==6.2.0
psycopg2==2.7.4 --no-binary psycopg2

# Django
django==3.2.4 

# Environment
django-environ==0.4.5

# Django REST Framework
djangorestframework==3.12.4

# Django cors headers
django-cors-headers==3.7.0

# JWT
pyjwt>=1.7.1

# Passwords security
argon2-cffi>=18.3.0

# Static files
whitenoise==4.1.2